Move components into their own directories
