+++ components/Filter/package.json
@@ -1,0 +1,6 @@
+{
+  "name": "Filter",
+  "version": "0.0.0",
+  "private": true,
+  "main": "./Filter.jsx"
+}
