+++ components/TaskList/package.json
@@ -1,0 +1,6 @@
+{
+  "name": "TaskList",
+  "version": "0.0.0",
+  "private": true,
+  "main": "./TaskList.jsx"
+}
